+++ server/src/main/java/com/seollem/server/member/controller/MemberController.java
@@ -11,32 +11,13 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/members")
 public class MemberController {
 
-    private final MemberRepository memberRepository;
-    private final BCryptPasswordEncoder bCryptPasswordEncoder;
-
-    public MemberController(MemberRepository memberRepository, BCryptPasswordEncoder bCryptPasswordEncoder) {
-        this.memberRepository = memberRepository;
-        this.bCryptPasswordEncoder = bCryptPasswordEncoder;
-    }
-
-    @GetMapping
-    public ResponseEntity<String> hello(){
-        return ResponseEntity.ok("hello");
-    }
-
-    @PostMapping("/join")
-    public String join(@RequestBody Member member) {
-        member.setPassword(bCryptPasswordEncoder.encode(member.getPassword()));
-        member.setRoles("ROLE_USER");
-        memberRepository.save(member);
-        return "회원 가입 완료";
-    }
 
     @GetMapping("/user")
     public String user() {
         return "user";
     }
-    // 추가
+
+
     @GetMapping("/admin")
     public String admin() {
         return "admin";
