+++ server/src/main/java/com/seollem/server/member/entity/Member.java
@@ -22,6 +22,7 @@ public class Member {
     private String name;
 
     private String password;
+
     private String roles;
